integrate drop log gui and drop log class

+++ inc/Functions.ahk
@@ -12,10 +12,26 @@ OnWM_LBUTTONDOWN(wParam, lParam, msg, hWnd) {
         return
     }
 
+    If !(dropLog.TripActive()) {
+        tooltip No trip started!
+        SetTimer, disableTooltip, -250
+        return
+    }
+
     id := SubStr(OutputAssociatedVar, InStr(OutputAssociatedVar, "#") + 1)
     obj := dropTable.GetDrop(id)
-    
-    tooltip, % obj.itemQuantity " x " obj.itemName
+    Obj.Delete("itemHighAlch")
+    Obj.Delete("itemImage")
+    Obj.Delete("itemPrice")
+    Obj.Delete("itemRarity")
+
+    g_selectedDrops.push(obj)
+
+    loop % g_selectedDrops.length()
+        drops .= g_selectedDrops[A_Index].itemQuantity " x " g_selectedDrops[A_Index].itemName ", "
+    drops := RTrim(drops, ", ")
+
+    logGui.SetText("edit1", drops)
 }
 
 ObjFullyClone(obj)
